cc: Make LayerTreeImpl dump layer list to FrameViewer and Devtools
As compositor will no longer know tree hierarchy information, it is
expected to dump a layer list instead of a layer tree to chrome tracing.

The value root_layer is also deleted because it is used to judge whether
the dumped layer_tree_impl has a list or not on the FrameViewer side.

Chrome://tracing will show a layer_tree_impl without depth, this is
expected.

This CL is blocked by Issue 1947553002, and potentially influenced by
the LayerTreeImpl's JSON issue.

Devtools need to deal with the new format information. On the devtools
side, we make the first drawsContent layer "contentRoot" and treat all
other layers as the contentRoot's child.

BUG=605940
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review-Url: https://codereview.chromium.org/1945813002
Cr-Commit-Position: refs/heads/master@{#397744}