Fix a bug with the Chrome legacy window at times becoming a top level window with a non functional button in the taskbar.
This bug typically occurs if the legacy window stays around after it is
removed from the original root window.

It looks like a recent change (https://codereview.chromium.org/1996163002)
which ensures that the legacy window is always created exacerbates this
problem by creating the legacy windows early and some of them stay around
after a tab switch. Typical cases include opening a pdf file in chrome
which causes at least two RWVHA instances to be active at any given point
and the legacy window which belongs to the original instance becomes a
button on the taskbar when the tab is switched.

Fix is to never reparent to the desktop window and instead always reparent
to the hidden window.

BUG=616410

Review-Url: https://codereview.chromium.org/2031213002
Cr-Commit-Position: refs/heads/master@{#397790}