Add smoke test for UsbService::GetDevices.
This test initializes the USB service and calls GetDevices() to exercise
the native USB device enumeration path on all platforms. Since we cannot
guarantee that a particular piece of hardware is connected the result
of this call is not checked. Tested enabled by --enable-gadget-tests
provide deeper converage when matching hardware is available.

BUG=None

Review-Url: https://codereview.chromium.org/2032303003
Cr-Commit-Position: refs/heads/master@{#397850}