Remove FilePath usage from the CrashReporterClient interface on Windows.
The CrashReporterClient implementation on Windows which is implemented by
the ChromeCrashReporterClient class will eventually move into chrome_elf
which loads very early in the chrome process. This requires it to not
depend on anything other than kernel32.dll and advapi32.dll.

Depending on base::FilePath pulls in FileUtil which on Windows brings with
it dependencies on message loop etc which won't work in chrome_elf.
Additionally the ChromeCrashReporterClient class currently depends on
content for constants, etc.

I defined these constants in the install_static library along with TODOs
to unify them. I plan to do that after finishing up the work for
chrome_elf.

The changes in this patch are as below:-
1. Remove FilePath from the CrashReporterClient functions on Windows.
2. Use the constants from install_static instead of those from content
   and base::Env.
3. Remove usage of base::Version from ChromeCrashReporterClient as that
   pulls in base logging which is not good for us.
4. Replace DCHECKs with asserts in ChromeCrashReporterClient

BUG=604923

Review-Url: https://codereview.chromium.org/2031833002
Cr-Commit-Position: refs/heads/master@{#397865}