[Layout API] Migrate TextAutosizer to block layout API
There are several references to document->layoutView() which can be
converted directly to document->layoutViewItem(), but there is one
that needs to use the block layout API shim introduced in
<https://codereview.chromium.org/2032023003>. It passes the layout
object around through several functions and uses a number of methods
that are not part of the API yet. As this file is already in core/layout
and I'm only converting enough to get rid of the document->layoutView
method, I see no reason to pollute the API with a raft of new methods.

BUG=499321

Review-Url: https://codereview.chromium.org/2041103002
Cr-Commit-Position: refs/heads/master@{#398305}