Support for bundling Chromium unittests into a tarball.
Without this change only the engine binary can be bundled in this way.

This change allows all the required files to successfully run in a
docker container can be bundled together in a tarball and uploaded to a
new location. This makes the process of exporting tests this way easily
repeatable and automatable.

This builds on the manifest generation work of
https://codereview.chromium.org/1997493002/

BUG=613000,608487,608483

Review-Url: https://codereview.chromium.org/2028353003
Cr-Commit-Position: refs/heads/master@{#398324}