Make field trials activated in the GPU process be reflected in browser.
This allows field trials that are queried in the GPU process be
reflected in UMA and crash reports. Currently, they are being
omitted.

This follows https://codereview.chromium.org/2020413002 - which
refactored this functionality out of the renderer process code, so that
it could be re-used (i.e. in this CL). One modification is made to that
refactoring, which is removing the sending of the IPC message from
the syncer class, since GPU and renderer processes require different
IPC messages.

BUG=616171

Review-Url: https://codereview.chromium.org/2034933002
Cr-Commit-Position: refs/heads/master@{#398345}