Introduce WebInsecureRequestPolicy.
This is the first step towards moving some of 'Upgrade-Insecure-Requests'
enforcement up to the browser process. It introduces WebInsecureRequestPolicy
which aims to represent both 'Upgrade-Insecure-Requests' and
'Block-All-Mixed-Content' as a single bitfield, as they're really two sides
of the same coin.

This patch uses the new type inside of //core/frame/csp. The next patch
will extend usage to 'SecurityContext' (which has a number of dependencies
including 'DocumentInit', 'FrameLoader' and etc, which makes it appealing
to split out). That will be followed by a patch which replicates the new
data to remote frames.

With that infrastructure in place, upgrading requests during redirects should
be (somewhat) straightforward.

[Step 1]: This patch.
[Step 2]: https://codereview.chromium.org/2040133003
[Step 3]: https://codereview.chromium.org/2046733003

BUG=617947
R=yoav@yoav.ws

Review-Url: https://codereview.chromium.org/2046523005
Cr-Commit-Position: refs/heads/master@{#398497}