exo: Avoid cross-fade animations for window without a configure callback.
When transitioning in/out of maximized or fullscreen mode we need to
make sure we have a configure callback before we allow the default
cross-fade animations. The configure callback provides a mechanism for
the client to inform us that a frame has taken the state change into
account and without this cross-fade animations are unreliable.

BUG=549781
TEST=exo_unittests

Review-Url: https://codereview.chromium.org/2049623002
Cr-Commit-Position: refs/heads/master@{#398616}