PlzNavigate: fix WebContentsImplTest.DirectNavigationToViewSourceWebUI
This CL fixes WebContentsImplTest.DirectNavigationToViewSourceWebUI. The test
in its current form only works because it is missing a call to
DidStartProvsionalLoad, which means that no handle is created for the
navigation. On commit, the pending NavigationEntry is reset, and a new one is
created with no virtual url and a rewritten url used for the title. In
PlzNavigate, we have a handle, so the pending entry with the viewsource virtual
url is kept, and the virtual url is displayed as a title. The CL fixes that by
adding the missing DidStartProvsionalLoad and updating the title expectations.

BUG=439423

Review-Url: https://codereview.chromium.org/2046843004
Cr-Commit-Position: refs/heads/master@{#398722}