Fixed callers that modify lists from getPropertyCSSValue()
Fixed callers that get a CSSValueList from getPropertyCSSValue(), then
modify the list before saving it back to the style. This is an actual
behavior change to the more correct behavior (copy a read-only version
of the list out of the style, modify it, then save it back to the
style).

This is pre-work for making StylePropertySet::getPropertyCSSValue return
a const CSSValue*, which will force all callers to implement this
behavior.

BUG=526586

Review-Url: https://codereview.chromium.org/2038373002
Cr-Commit-Position: refs/heads/master@{#398829}