[css-grid] Fix a crash when inspecting grids with abs. positioned children
Absolute positioning occurs after layout of the grid and its in-flow
contents, and does not contribute to the sizing of any grid tracks or affect
the size/configuration of the grid in any way. This means that we should
treat as empty any grid whose only children are absolutely positioned items.

Since refs/heads/master@{#396192} empty grids are no longer internally
represented by a 1x1 matrix. As we were not considering
grids-with-only-absolutely-positioned-children as empty, we were trying to
access some invalid position in the internal representation of the grid
triggering a crash (or ASSERT in Debug) in the Vector code.

Review-Url: https://codereview.chromium.org/2049503005
Cr-Commit-Position: refs/heads/master@{#398939}