Decouple capture timestamp and reference time
In this CL, we make the captured VideoFrame to take capture device
timestamp; and take system clock reference time.

Background:
Previously, we have only timestamp parameter, and it actually
plays the role as timestamp and reference_time.
Here we define timestamp as the ideal time span between
the current frame and the first frame in the stream.
Define reference time as the system clock time when the frame
gets generated.

Ideally, timestamp can be derived from reference_time.
However, in reality, especially in a multi process application
like Chrome, system clock read has noise on the order of 5ms.
To overcome that, we found the original timestamp from camera
driver is much smoother. But the defect is that the camera
clock would have a drift with respect to the system clock.

Thus a perfect solution would be that we take system time for
reference time, which is used for AV sync; we stream timestamp
to the remote side, which will make the rendering smooth.

This CL is complete for the complete fix.
BUG=601657
TBR=palmer@chromium.org

Review-Url: https://codereview.chromium.org/2045813003
Cr-Commit-Position: refs/heads/master@{#399304}